requirements.txt
google-generativeai
protobuf
z3-solver
pdoc3
pipreqs
